pyproject.toml
[project]
name = "coretext-interview-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.52.0",
    "fastapi[all]>=0.115.12",
    "fastmcp>=2.6.1",
    "langchain-anthropic>=0.3.13",
    "langchain-community>=0.3.24",
    "langchain-mcp-adapters>=0.1.1",
    "langgraph>=0.4.7",
    "mcp>=1.9.1",
    "python-dotenv>=1.1.0",
    "sse-starlette>=2.3.5",
]
